Removed plan.txt Refactor abstract workflow decorator to reuse wf and update docstrings Changed name ModuleNode to Module and dag.py to core
